api/utils.py
import mysql.connector
from pymongo import MongoClient
from cassandra.cluster import Cluster
from cassandra.auth import PlainTextAuthProvider
from django.conf import settings

def get_mysql_connection(connection):
    return mysql.connector.connect(
        host=connection.host,
        port=connection.port,
        user=connection.username,
        password=connection.password,
        database=connection.database_name
    )

def get_mongodb_connection():
    client = MongoClient(settings.MONGODB_URI)
    return client[settings.MONGODB_NAME]

def get_cassandra_connection():
    auth_provider = PlainTextAuthProvider(username=settings.CASSANDRA_USERNAME, password=settings.CASSANDRA_PASSWORD)
    cluster = Cluster(settings.CASSANDRA_HOSTS, auth_provider=auth_provider)
    return cluster.connect(settings.CASSANDRA_KEYSPACE)

def migrate_data(job):
    source_conn = get_connection(job.source)
    dest_conn = get_connection(job.destination)

    # Implement the migration logic here
    # This is a placeholder and needs to be implemented based on the specific requirements

    job.status = 'completed'
    job.save()

def execute_stored_procedure(connection, procedure_name, parameters):
    if connection.db_type != 'mysql':
        raise ValueError("Stored procedures are only supported for MySQL")

    mysql_conn = get_mysql_connection(connection)
    cursor = mysql_conn.cursor()

    try:
        cursor.callproc(procedure_name, parameters)
        results = []
        for result in cursor.stored_results():
            results.extend(result.fetchall())
        mysql_conn.commit()
        return results
    finally:
        cursor.close()
        mysql_conn.close()

def get_connection(db_connection):
    if db_connection.db_type == 'mysql':
        return get_mysql_connection(db_connection)
    elif db_connection.db_type == 'mongodb':
        return get_mongodb_connection()
    elif db_connection.db_type == 'cassandra':
        return get_cassandra_connection()
    else:
        raise ValueError(f"Unsupported database type: {db_connection.db_type}")

