requirements.txt
Django==3.2.9
djangorestframework==3.12.4
django-cors-headers==3.10.0
mysql-connector-python==8.0.27
pymongo==3.12.1
cassandra-driver==3.25.0

